+++ web/modules/custom/hoeringsportal_project/modules/hoeringsportal_project_fixtures/src/Fixture/ProjectMainPageFixture.php
@@ -1,0 +1,116 @@
+<?php
+
+namespace Drupal\hoeringsportal_project_fixtures\Fixture;
+
+use Drupal\content_fixtures\Fixture\AbstractFixture;
+use Drupal\content_fixtures\Fixture\DependentFixtureInterface;
+use Drupal\content_fixtures\Fixture\FixtureGroupInterface;
+use Drupal\hoeringsportal_base_fixtures\Fixture\MediaFixture;
+use Drupal\hoeringsportal_base_fixtures\Fixture\ParagraphFixture;
+use Drupal\node\Entity\Node;
+use Drupal\paragraphs\Entity\Paragraph;
+
+/**
+ * Landing page fixture.
+ *
+ * @package Drupal\hoeringsportal_project_fixtures\Fixture
+ */
+class ProjectMainPageFixture extends AbstractFixture implements DependentFixtureInterface, FixtureGroupInterface {
+
+  /**
+   * {@inheritdoc}
+   */
+  public function load() {
+    // Project main page.
+    $entity = Node
+      ::create([
+        'type' => 'project_main_page',
+      ])
+        ->setTitle('Project')
+        ->set('field_project_category', [
+          $this->getReference('project_categories:Byudvikling'),
+        ])
+        ->set('field_short_description', 'This is the first project')
+        ->set('field_project_image', [
+        ['target_id' => $this->getReference('media:Medium1')->id()],
+        ]);
+
+    $paragraph = Paragraph
+      ::create([
+        'type' => 'image',
+      ])
+        ->set('field_paragraph_image', [
+        ['target_id' => $this->getReference('media:Medium2')->id()],
+        ]);
+    $paragraph->save();
+    $entity->field_content_sections->appendItem([
+      'target_id' => $paragraph->id(),
+      'target_revision_id' => $paragraph->getRevisionId(),
+    ]);
+
+    $paragraph = Paragraph
+      ::create([
+        'type' => 'info_box',
+      ])
+        ->set('field_paragraph_title', 'Important!')
+        ->set('field_content_block_text', [
+          'value' => <<<'BODY'
+<p>Beware that this is the <em>first</em> project.</p>
+BODY,
+          'format' => 'filtered_html',
+        ]);
+    $paragraph->save();
+    $entity->field_content_sections->appendItem([
+      'target_id' => $paragraph->id(),
+      'target_revision_id' => $paragraph->getRevisionId(),
+    ]);
+
+    $entity->save();
+    $this->addReference('node:project_main_page:1', $entity);
+
+    $entity = $entity->createDuplicate();
+    $entity
+      ->setTitle('Another project')
+      ->set('field_short_description', 'We have more than one project')
+      ->set('field_project_image', [
+        ['target_id' => $this->getReference('media:Map1')->id()],
+      ]);
+
+    $entity->save();
+    $this->addReference('node:project_main_page:2', $entity);
+
+    $entity = $entity->createDuplicate();
+    $entity
+      ->setTitle('The third project')
+      ->set('field_project_category', [
+        $this->getReference('project_categories:Offentlig transport'),
+      ])
+      ->set('field_short_description', 'We have more than one project')
+      ->set('field_project_image', [
+        ['target_id' => $this->getReference('media:Map1')->id()],
+      ]);
+
+    $entity->save();
+  }
+
+  /**
+   * {@inheritdoc}
+   */
+  public function getDependencies() {
+    return [
+      ProjectCategoryFixture::class,
+      BlockContentFixture::class,
+      MediaFixture::class,
+      ParagraphFixture::class,
+      ProjectPageFixture::class,
+    ];
+  }
+
+  /**
+   * {@inheritdoc}
+   */
+  public function getGroups() {
+    return ['node', 'project'];
+  }
+
+}